services/asr/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
openai-whisper==20231117
# faster-whisper==1.0.0  # Uncomment for faster inference
torch==2.1.1
torchaudio==2.1.1
numpy==1.24.3